Move history models into policy module
