+++ src/dr_wandb/export/__init__.py
@@ -6,12 +6,15 @@ from dr_wandb.export.models import (
     ExportState,
     ExportSummary,
     FetchMode,
+    RunSnapshot,
+)
+from dr_wandb.export.policy import (
+    HistoryPolicy,
     HistoryPolicyContext,
     HistoryRow,
     HistoryWindow,
-    RunSnapshot,
+    StaticHistoryPolicy,
 )
-from dr_wandb.export.policy import HistoryPolicy, StaticHistoryPolicy
 from dr_wandb.export.record_store import RecordStore
 
 __all__ = [
